Separated out address code in Wallet plugin
