+++ package.json
@@ -30,7 +30,7 @@
     "sia-repo": "go get -u github.com/NebulousLabs/Sia/...",
     "sia-import": "(test -d Sia || mkdir -p Sia) && (cp $GOPATH/bin/siad Sia & cp $GOPATH/bin/siac Sia)",
     "sia-kill": "killall siad; while pgrep siad; do sleep .5; done;",
-    "sia-reimport": "npm run sia-repo && npm run sia-kill && npm run sia-import",
+    "sia-refresh": "npm run sia-kill && rm -rf Sia && npm run sia-import",
     "release-args": "electron-packager . Sia-UI --app-version=0.4.8-beta --version=0.34.1 --overwrite --prune --ignore=release --out=release/v0.4.8-beta --icon=app/assets/icon",
     "release": "npm run sia-import ; npm run release-args -- --all && ./.finalizeRelease.sh",
     "release-lin64": "npm run release-args -- --platform=linux --arch=x64 && ./.finalizeRelease.sh",
