+++ plugins/Wallet/index.html
@@ -154,6 +154,7 @@
 		<script src='js/helpers.js'></script>
 		<script src='js/locking.js'></script>
 		<script src='js/buttons.js'></script>
+		<script src='js/addresses.js'></script>
 		<script src='js/lifecycle.js'></script>
 	</body>
 </html>
